RAG Chat with Multiple Docs using Llama Index/requirements.txt
llama-index
openai
pypdf
python-dotenv
streamlit
google-generativeai